+++ src/args.rs
@@ -376,7 +376,10 @@ pub(crate) enum Generate {
         /// Output as JSON instead of human-readable text.
         #[arg(short, long)]
         json: bool,
-        /// Filename to write the TOML to, defaults to `server.toml`.
+        /// Generate a single-party config that allows all exec activities and outbound HTTP.
+        #[arg(long)]
+        trusted: bool,
+        /// Filename to write the TOML to, defaults to <stdout>.
         output: Option<PathBuf>,
         /// Overwrite existing file.
         #[arg(long, short)]
@@ -387,7 +390,7 @@ pub(crate) enum Generate {
         /// Output as JSON instead of human-readable text.
         #[arg(short, long)]
         json: bool,
-        /// Filename to write the TOML to, defaults to `deployment.toml`.
+        /// Filename to write the TOML to, defaults to <stdout>.
         output: Option<PathBuf>,
         /// Overwrite existing file.
         #[arg(long, short)]